gocr: Fix problem with patch whitespace
The makefile patch contains some trailing whitespace that must be present for the patch to be applied correctly. By including the patch in the formula as DATA, the whitespace kept being automatically removed by editors trying to be helpful. To stop this happening, this commit moves the patch to an external source.